+++ node/components.js
@@ -1,4 +1,338 @@
-function sendHiMessage(recipientId) {
+const
+  crypto = require('crypto'),
+  request = require('request'),
+  config = require('config');
+
+// App Secret can be retrieved from the App Dashboard
+export const APP_SECRET = (process.env.MESSENGER_APP_SECRET) ?
+  process.env.MESSENGER_APP_SECRET :
+  config.get('appSecret');
+
+// Generate a page access token for your page from the App Dashboard
+export const PAGE_ACCESS_TOKEN = (process.env.MESSENGER_PAGE_ACCESS_TOKEN) ?
+  (process.env.MESSENGER_PAGE_ACCESS_TOKEN) :
+  config.get('pageAccessToken');
+
+export function verifyRequestSignature(req, res, buf) {
+  var signature = req.headers["x-hub-signature"];
+
+  if (!signature) {
+    // For testing, let's log an error. In production, you should throw an
+    // error.
+    console.error("Couldn't validate the signature.");
+  } else {
+    var elements = signature.split('=');
+    var method = elements[0];
+    var signatureHash = elements[1];
+
+    var expectedHash = crypto.createHmac('sha1', APP_SECRET)
+                        .update(buf)
+                        .digest('hex');
+
+    if (signatureHash != expectedHash) {
+      throw new Error("Couldn't validate the request signature.");
+    }
+  }
+}
+
+/*
+ * Authorization Event
+ *
+ * The value for 'optin.ref' is defined in the entry point. For the "Send to
+ * Messenger" plugin, it is the 'data-ref' field. Read more at
+ * https://developers.facebook.com/docs/messenger-platform/webhook-reference/authentication
+ *
+ */
+export function receivedAuthentication(event) {
+  var senderID = event.sender.id;
+  var recipientID = event.recipient.id;
+  var timeOfAuth = event.timestamp;
+
+  // The 'ref' field is set in the 'Send to Messenger' plugin, in the 'data-ref'
+  // The developer can set this to an arbitrary value to associate the
+  // authentication callback with the 'Send to Messenger' click event. This is
+  // a way to do account linking when the user clicks the 'Send to Messenger'
+  // plugin.
+  var passThroughParam = event.optin.ref;
+
+  console.log("Received authentication for user %d and page %d with pass " +
+    "through param '%s' at %d", senderID, recipientID, passThroughParam,
+    timeOfAuth);
+
+  // When an authentication is received, we'll send a message back to the sender
+  // to let them know it was successful.
+  sendTextMessage(senderID, "Authentication successful");
+}
+
+/*
+ * Message Event
+ *
+ * This event is called when a message is sent to your page. The 'message'
+ * object format can vary depending on the kind of message that was received.
+ * Read more at https://developers.facebook.com/docs/messenger-platform/webhook-reference/message-received
+ *
+ * For this example, we're going to echo any text that we get. If we get some
+ * special keywords ('button', 'generic', 'receipt'), then we'll send back
+ * examples of those bubbles to illustrate the special message bubbles we've
+ * created. If we receive a message with an attachment (image, video, audio),
+ * then we'll simply confirm that we've received the attachment.
+ *
+ */
+export function receivedMessage(event) {
+  var senderID = event.sender.id;
+  var recipientID = event.recipient.id;
+  var timeOfMessage = event.timestamp;
+  var message = event.message;
+
+  console.log("Received message for user %d and page %d at %d with message:",
+    senderID, recipientID, timeOfMessage);
+  console.log(JSON.stringify(message));
+
+  var isEcho = message.is_echo;
+  var messageId = message.mid;
+  var appId = message.app_id;
+  var metadata = message.metadata;
+
+  // You may get a text or attachment but not both
+  var messageText = message.text;
+  var messageAttachments = message.attachments;
+  var quickReply = message.quick_reply;
+
+  if (isEcho) {
+    // Just logging message echoes to console
+    console.log("Received echo for message %s and app %d with metadata %s",
+      messageId, appId, metadata);
+    return;
+  } else if (quickReply) {
+    var quickReplyPayload = quickReply.payload;
+    console.log("Quick reply for message %s with payload %s",
+      messageId, quickReplyPayload);
+
+    sendTextMessage(senderID, "Quick reply tapped");
+    return;
+  }
+
+  if (messageText) {
+
+    // If we receive a text message, check to see if it matches any special
+    // keywords and send back the corresponding example. Otherwise, just echo
+    // the text we received.
+    switch (messageText.replace(/[^\w\s]/gi, '').trim().toLowerCase()) {
+      case 'hello':
+      case 'hi':
+        sendHiMessage(senderID);
+        break;
+
+      case 'image':
+        requiresServerURL(sendImageMessage, [senderID]);
+        break;
+
+      case 'gif':
+        requiresServerURL(sendGifMessage, [senderID]);
+        break;
+
+      case 'audio':
+        requiresServerURL(sendAudioMessage, [senderID]);
+        break;
+
+      case 'video':
+        requiresServerURL(sendVideoMessage, [senderID]);
+        break;
+
+      case 'file':
+        requiresServerURL(sendFileMessage, [senderID]);
+        break;
+
+      case 'button':
+        sendButtonMessage(senderID);
+        break;
+
+      case 'generic':
+        requiresServerURL(sendGenericMessage, [senderID]);
+        break;
+
+      case 'receipt':
+        requiresServerURL(sendReceiptMessage, [senderID]);
+        break;
+
+      case 'quick reply':
+        sendQuickReply(senderID);
+        break;
+
+      case 'read receipt':
+        sendReadReceipt(senderID);
+        break;
+
+      case 'typing on':
+        sendTypingOn(senderID);
+        break;
+
+      case 'typing off':
+        sendTypingOff(senderID);
+        break;
+
+      case 'account linking':
+        requiresServerURL(sendAccountLinking, [senderID]);
+        break;
+
+      default:
+        sendTextMessage(senderID, messageText);
+    }
+  } else if (messageAttachments) {
+    sendTextMessage(senderID, "Message with attachment received");
+  }
+}
+
+
+/*
+ * Delivery Confirmation Event
+ *
+ * This event is sent to confirm the delivery of a message. Read more about
+ * these fields at https://developers.facebook.com/docs/messenger-platform/webhook-reference/message-delivered
+ *
+ */
+export function receivedDeliveryConfirmation(event) {
+  var senderID = event.sender.id;
+  var recipientID = event.recipient.id;
+  var delivery = event.delivery;
+  var messageIDs = delivery.mids;
+  var watermark = delivery.watermark;
+  var sequenceNumber = delivery.seq;
+
+  if (messageIDs) {
+    messageIDs.forEach(function(messageID) {
+      console.log("Received delivery confirmation for message ID: %s",
+        messageID);
+    });
+  }
+
+  console.log("All message before %d were delivered.", watermark);
+}
+
+
+/*
+ * Postback Event
+ *
+ * This event is called when a postback is tapped on a Structured Message.
+ * https://developers.facebook.com/docs/messenger-platform/webhook-reference/postback-received
+ *
+ */
+export function receivedPostback(event) {
+  var senderID = event.sender.id;
+  var recipientID = event.recipient.id;
+  var timeOfPostback = event.timestamp;
+
+  // The 'payload' param is a developer-defined field which is set in a postback
+  // button for Structured Messages.
+  var payload = event.postback.payload;
+
+  console.log("Received postback for user %d and page %d with payload '%s' " +
+    "at %d", senderID, recipientID, payload, timeOfPostback);
+
+  // When a postback is called, we'll send a message back to the sender to
+  // let them know it was successful
+  sendTextMessage(senderID, "Postback called");
+}
+
+/*
+ * Message Read Event
+ *
+ * This event is called when a previously-sent message has been read.
+ * https://developers.facebook.com/docs/messenger-platform/webhook-reference/message-read
+ *
+ */
+export function receivedMessageRead(event) {
+  var senderID = event.sender.id;
+  var recipientID = event.recipient.id;
+
+  // All messages before watermark (a timestamp) or sequence have been seen.
+  var watermark = event.read.watermark;
+  var sequenceNumber = event.read.seq;
+
+  console.log("Received message read event for watermark %d and sequence " +
+    "number %d", watermark, sequenceNumber);
+}
+
+/*
+ * Account Link Event
+ *
+ * This event is called when the Link Account or UnLink Account action has been
+ * tapped.
+ * https://developers.facebook.com/docs/messenger-platform/webhook-reference/account-linking
+ *
+ */
+export function receivedAccountLink(event) {
+  var senderID = event.sender.id;
+  var recipientID = event.recipient.id;
+
+  var status = event.account_linking.status;
+  var authCode = event.account_linking.authorization_code;
+
+  console.log("Received account link event with for user %d with status %s " +
+    "and auth code %s ", senderID, status, authCode);
+}
+
+/*
+ * If users came here through testdrive, they need to configure the server URL
+ * in default.json before they can access local resources likes images/videos.
+ */
+export function requiresServerURL(next, [recipientId, ...args]) {
+  if (SERVER_URL === "to_be_set_manually") {
+    var messageData = {
+      recipient: {
+        id: recipientId
+      },
+      message: {
+        text: `
+We have static resources like images and videos available to test, but you need to update the code you downloaded earlier to tell us your current server url.
+1. Stop your node server by typing ctrl-c
+2. Paste the result you got from running "lt —port 5000" into your config/default.json file as the "serverURL".
+3. Re-run "node app.js"
+Once you've finished these steps, try typing “video” or “image”.
+        `
+      }
+    }
+
+    callSendAPI(messageData);
+  } else {
+    next.apply(this, [recipientId, ...args]);
+  }
+}
+
+
+
+/*
+ * Call the Send API. The message data goes in the body. If successful, we'll
+ * get the message id in a response
+ *
+ */
+export function callSendAPI(messageData) {
+  request({
+    uri: 'https://graph.facebook.com/v2.6/me/messages',
+    qs: { access_token: PAGE_ACCESS_TOKEN },
+    method: 'POST',
+    json: messageData
+
+  }, function (error, response, body) {
+    if (!error && response.statusCode == 200) {
+      var recipientId = body.recipient_id;
+      var messageId = body.message_id;
+
+      if (messageId) {
+        console.log("Successfully sent message with id %s to recipient %s",
+          messageId, recipientId);
+      } else {
+      console.log("Successfully called Send API for recipient %s",
+        recipientId);
+      }
+    } else {
+      console.error("Failed calling Send API", response.statusCode, response.statusMessage, body.error);
+    }
+  });
+}
+
+
+export function sendHiMessage(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -13,14 +347,15 @@ For more details on how to create commands, go to https://developers.facebook.co
       `
     }
   }
-  return messageData;
+
+  callSendAPI(messageData);
 }
 
 /*
  * Send an image using the Send API.
  *
  */
-function sendImageMessage(recipientId) {
+export function sendImageMessage(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -42,7 +377,7 @@ function sendImageMessage(recipientId) {
  * Send a Gif using the Send API.
  *
  */
-function sendGifMessage(recipientId) {
+export function sendGifMessage(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -64,7 +399,7 @@ function sendGifMessage(recipientId) {
  * Send audio using the Send API.
  *
  */
-function sendAudioMessage(recipientId) {
+export function sendAudioMessage(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -86,7 +421,7 @@ function sendAudioMessage(recipientId) {
  * Send a video using the Send API.
  *
  */
-function sendVideoMessage(recipientId) {
+export function sendVideoMessage(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -108,7 +443,7 @@ function sendVideoMessage(recipientId) {
  * Send a file using the Send API.
  *
  */
-function sendFileMessage(recipientId) {
+export function sendFileMessage(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -127,10 +462,28 @@ function sendFileMessage(recipientId) {
 }
 
 /*
+ * Send a text message using the Send API.
+ *
+ */
+export function sendTextMessage(recipientId, messageText) {
+  var messageData = {
+    recipient: {
+      id: recipientId
+    },
+    message: {
+      text: messageText,
+      metadata: "DEVELOPER_DEFINED_METADATA"
+    }
+  };
+
+  callSendAPI(messageData);
+}
+
+/*
  * Send a button message using the Send API.
  *
  */
-function sendButtonMessage(recipientId) {
+export function sendButtonMessage(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -166,7 +519,7 @@ function sendButtonMessage(recipientId) {
  * Send a Structured Message (Generic Message type) using the Send API.
  *
  */
-function sendGenericMessage(recipientId) {
+export function sendGenericMessage(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -217,7 +570,7 @@ function sendGenericMessage(recipientId) {
  * Send a receipt message using the Send API.
  *
  */
-function sendReceiptMessage(recipientId) {
+export function sendReceiptMessage(recipientId) {
   // Generate a random receipt ID as the API requires a unique ID
   var receiptId = "order" + Math.floor(Math.random()*1000);
 
@@ -283,7 +636,7 @@ function sendReceiptMessage(recipientId) {
  * Send a message with Quick Reply buttons.
  *
  */
-function sendQuickReply(recipientId) {
+export function sendQuickReply(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -317,7 +670,7 @@ function sendQuickReply(recipientId) {
  * Send a read receipt to indicate the message has been read
  *
  */
-function sendReadReceipt(recipientId) {
+export function sendReadReceipt(recipientId) {
   console.log("Sending a read receipt to mark message as seen");
 
   var messageData = {
@@ -334,7 +687,7 @@ function sendReadReceipt(recipientId) {
  * Turn typing indicator on
  *
  */
-function sendTypingOn(recipientId) {
+export function sendTypingOn(recipientId) {
   console.log("Turning typing indicator on");
 
   var messageData = {
@@ -351,7 +704,7 @@ function sendTypingOn(recipientId) {
  * Turn typing indicator off
  *
  */
-function sendTypingOff(recipientId) {
+export function sendTypingOff(recipientId) {
   console.log("Turning typing indicator off");
 
   var messageData = {
@@ -368,7 +721,7 @@ function sendTypingOff(recipientId) {
  * Send a message with the account linking call-to-action
  *
  */
-function sendAccountLinking(recipientId) {
+export function sendAccountLinking(recipientId) {
   var messageData = {
     recipient: {
       id: recipientId
@@ -391,6 +744,32 @@ function sendAccountLinking(recipientId) {
   callSendAPI(messageData);
 }
 
-
-
-module.exports.sendHiMessage = sendHiMessage
+  /**
+module.exports = {
+  APP_SECRET: APP_SECRET,
+  sendHiMessage: sendHiMessage,
+  verifyRequestSignature: verifyRequestSignature,
+  receivedAuthentication: receivedAuthentication,
+  receivedMessage: receivedMessage,
+  receivedDeliveryConfirmation: receivedDeliveryConfirmation,
+  receivedPostback: receivedPostback,
+  receivedMessageRead: receivedMessageRead,
+  receivedAccountLink: receivedAccountLink,
+  requiresServerURL: requiresServerURL,
+  callSendAPI: callSendAPI,
+  sendHiMessage: sendHiMessage,
+  sendGifMessage: sendGifMessage,
+  sendAudioMessage: sendAudioMessage,
+  sendVideoMessage: sendVideoMessage,
+  sendFileMessage: sendFileMessage,
+  sendTextMessage: sendTextMessage,
+  sendButtonMessage: sendButtonMessage,
+  sendGenericMessage: sendGenericMessage,
+  sendReceiptMessage: sendReceiptMessage,
+  sendQuickReply: sendQuickReply, 
+  sendReadReceipt: sendReadReceipt,
+  sendTypingOn: sendTypingOn,
+  sendTypingOff: sendTypingOff,
+  sendAccountLinking: sendAccountLinking
+}
+*/
