+++ node/package.json
@@ -18,8 +18,8 @@
     "body-parser": "^1.15.0",
     "config": "^1.20.4",
     "ejs": "^2.4.2",
+    "esm": "^3.0.84",
     "express": "^4.13.4",
-    "localtunnel": "^1.9.1",
     "request": "^2.88.0"
   },
   "engines": {
